refactor: move updateLearnedRecipes to its own file
